feat: handle dropped and replaced transactions
Currently if a user speeds up a transaction the old transaction will report as failed and the new one will not exist in useDapp state. This change adds some handling around the failed transaction to check if it was replaced and repriced and should therefore continue on as a new pending transaction.