replace Object.fromEntries with Array.reduce
Rationale behind change:
Object.fromEntries needs to be polyfilled in any project using this right now
and the switch to Array.reduce removes this need. To me it makes sense not
to have dependencies like that. Sentry for the current project I work with is
filled with `Object.fromEntries is not a function` errors.